clean up post, add offer mode to posts, remove offer / reject buttons, begin form on offer
