--- src/components/pages/Acceptance.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-class AcceptancePage extends React.Component {
-    constructor() {
-        super();
-    }
-
-    render() {
-        return (
-            <div>
-                Hey!
-            </div>  
-        )
-    }
-}
