+++ src/components/pages/Landing.js
@@ -2,17 +2,17 @@ import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import { Post } from '../lists/PostList';
+import Post from '../cards/Post';
 import * as colors from '../../colors';
 import * as utils from '../../utils';
 import BadgeButton from '../buttons/BadgeButton';
 
-const FAKE_DESCRIPTION =`
+const FAKE_DESCRIPTION = `
 We're a relatively small
 indie game company based out of Seattle, WA looking for a youtuber to review our game.
 `
 
-const CARD_EXPLANATION =`
+const CARD_EXPLANATION = `
 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris augue ipsum, sodales vitae laoreet vitae, interdum vel risus. Nulla facilisi. Suspendisse posuere rhoncus placerat. Nunc ut ante eget odio auctor mattis ut sit amet augue. In tincidunt, sem at aliquam varius, risus sapien tempus dui, eu iaculis nisl lorem eu enim. Suspendisse venenatis nisi lacus. Aliquam tellus felis, pretium vitae bibendum sit amet, elementum ut enim. Pellentesque facilisis non dui eget porttitor. Nam a mattis sem, at congue ex. Ut lectus neque, laoreet eget lectus ut, venenatis tempus turpis.
 `
 
@@ -25,21 +25,21 @@ const LandingSignInButton = (props) => {
     <BadgeButton
       {...props}
       style={{
-        backgroundColor: colors.myMessage, 
-        padding: "12px", 
+        backgroundColor: colors.myMessage,
+        padding: "12px",
         color: "white",
         paddingRight: "24px",
         paddingLeft: "24px"
       }}
     >
-      <h1 
+      <h1
         style={{
-          borderBottom: "2px solid white", 
+          borderBottom: "2px solid white",
           paddingBottom: "4px",
           borderRadius: "2px"
         }}
-      > 
-        { props.title } 
+      >
+        {props.title}
       </h1>
     </BadgeButton>
   )
@@ -54,7 +54,7 @@ const LandingTitle = (props) => {
         textShadow: `2px -2px ${colors.yaDevPurple}`
       }}
     >
-      { props.children }
+      {props.children}
     </h1>
   )
 }
@@ -68,7 +68,7 @@ const LandingParagraph = (props) => {
         maxWidth: "60rem"
       }}
     >
-      { props.children }
+      {props.children}
     </p>
   )
 }
@@ -89,22 +89,22 @@ const Landing = (props) => {
             Easily find advertisement for your games
           </LandingTitle>
           <LandingParagraph>
-            { CARD_EXPLANATION }
+            {CARD_EXPLANATION}
           </LandingParagraph>
         </Col>
         <Col md={4}>
-            <Post
-              noHide
-              post={{
-                type: "developer",
-                price: 50,
-                description: FAKE_DESCRIPTION,
-                user: {
-                  avgRating: 4,
-                  completed: 31
-                }
-              }}
-            />
+          <Post
+            noHide
+            post={{
+              type: "developer",
+              price: 50,
+              description: FAKE_DESCRIPTION,
+              user: {
+                avgRating: 4,
+                completed: 31
+              }
+            }}
+          />
         </Col>
       </Row>
       <hr />
@@ -133,27 +133,27 @@ const Landing = (props) => {
             Not a developer?
           </LandingTitle>
           <LandingParagraph>
-              { CONTENT_CREATOR_EXPLANATION }
+            {CONTENT_CREATOR_EXPLANATION}
           </LandingParagraph>
         </Col>
       </Row>
       <Row>
-          <Col md={12} style={{textAlign: "center"}}>
+        <Col md={12} style={{ textAlign: "center" }}>
           <hr />
-            <LandingTitle>
-              Get started with Github or Google
+          <LandingTitle>
+            Get started with Github or Google
             </LandingTitle>
-            <hr />
-            <LandingSignInButton
-              title="Github"
-              onClick={utils.loginGithub}
-            />
-            <LandingSignInButton
-              title="Google"
-              onClick={utils.loginGoogle}
-            />
-          </Col>
-        </Row>
+          <hr />
+          <LandingSignInButton
+            title="Github"
+            onClick={utils.loginGithub}
+          />
+          <LandingSignInButton
+            title="Google"
+            onClick={utils.loginGoogle}
+          />
+        </Col>
+      </Row>
     </Container>
   )
 }
